+++ lib/features/notificacoes/domain/repositories/preferencias_notificacao_repository.dart
@@ -1,0 +1,7 @@
+import '../entities/preferencias_notificacao_entity.dart';
+
+abstract class PreferenciasNotificacaoRepository {
+  Future<PreferenciasNotificacaoEntity> obter();
+
+  Future<void> salvar(PreferenciasNotificacaoEntity preferencias);
+}