+++ lib/services/notificacoes/domain/repositories/notificacao_local_client.dart
@@ -1,6 +1,10 @@
 import '../entities/notificacao_local_request.dart';
 
 abstract class NotificacaoLocalClient {
+  Future<void> inicializar();
+
+  Future<bool> solicitarPermissao();
+
   Future<void> agendar(NotificacaoLocalRequest request);
 
   Future<void> cancelar(int id);